backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
pydantic[email]>=2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
anyio>=4.8.0,<5.0.0
email-validator>=2.0.0

